@@ -18,9 +18,9 @@
 
 package com.generalbytes.bitrafael.server.api;
 
-import javax.ws.rs.Path;
-import javax.ws.rs.Produces;
-import javax.ws.rs.core.MediaType;
+import jakarta.ws.rs.Path;
+import jakarta.ws.rs.Produces;
+import jakarta.ws.rs.core.MediaType;
 
 @Path("/dash/v1")
 @Produces(MediaType.APPLICATION_JSON)
